Fixed data source processing
The comparators weren't displaying since the conversion to using TYPE_JSARRAY because YUI DataSource syntax varies between TYPE_JSARRAY and TYPE_JSON, with only the latter supporting the dotted sub-element notation. Now we simply bite the bullet and use a FunctionDataSource so we can process the data however we want.