Experimental tree de-noising
Added a command-line option which tells webcompare to remove elements
matching a list of XPath expressions before doing any document
comparisons. This is handy when comparing sites which have changed
templates to the point that results will be heavily skewed due to the
text contained in headers, navigation, etc. even though the content is
quite similar.

Usage:

./webcompare.py --origin-noise-xpath-file=origin-noise.xpaths --target-noise-xpath-file=target-noise.xpaths